modules/restendpoint/requirements.txt
azure-iot-device~=2.7.0
Flask==2.1.1
paho-mqtt==1.6.1